// db/model_generated.ts
export namespace playersFields {
    export type id = string;
    export type player_first_name = string;
    export type player_last_name = string;
    export type player_email = string;
    export type player_phone = string;
    export type is_team_owner = boolean;

}

export interface players {
    id: playersFields.id;
    player_first_name: playersFields.player_first_name;
    player_last_name: playersFields.player_last_name;
    player_email: playersFields.player_email;
    player_phone: playersFields.player_phone;
    is_team_owner: playersFields.is_team_owner;

}